requirements.txt
# Core dependencies
streamlit>=1.28.0
langgraph>=0.0.60
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langchain-chroma>=0.1.0

# Vector database and embeddings
chromadb>=0.4.15
sentence-transformers>=2.2.2

# Additional utilities
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
markdown>=3.5.0
tiktoken>=0.5.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
